Move the prefill history test onto TestGenerate.
Uses the Qwen model the class already loads instead of a local dummy, so
the new file and its stub model go away.
